feat: add error pass-through
Adds error pass-through which adds a pass-through-flag to errors. A function named 'Finalize' checks for such a flag and unpacks the wrapped error or keeps the original one. This can be used when other libraries or modules depend on actual errors being passed, e.g., Kubernetes apierror. Errors can then be logged regularly but when they are handed over to the third-party caller, Finalize extracts the original one.

Signed-off-by: Lennart Altenhof <la@datatroniq.com>